+++ src/Components/Cards/Cards.css
@@ -10,10 +10,10 @@
   background-color: var(--bg-color);
 }
 
-.video-card-vertical:hover {
+.animate-card:hover {
   transition: 300ms cubic-bezier(0.175, 0.885, 0.32, 1.275);
-  transform: scale(1.1);
-  z-index: 2;
+  transform: scale(1.05);
+  z-index: 1;
 }
 
 .video-card-vertical .card-img-container {
