+++ src/Pages/Playlists/SinglePlaylist/Playlist.jsx
@@ -1,11 +1,31 @@
+import { useEffect } from "react";
+import { useParams } from "react-router-dom";
 import { VerticalCard } from "../../../Components/Cards/VerticalCard";
-import { useVideo } from "../../../Context";
+import { useAuth, useAxiosCalls, useVideo } from "../../../Context";
 import "./Playlist.css";
 
 export const Playlist = () => {
   const {
     videoState: { singlePlaylist },
   } = useVideo();
+
+  const { getPlayListFromServer } = useAxiosCalls();
+
+  const {
+    auth: { token },
+  } = useAuth();
+
+  const { playlistId } = useParams();
+
+  const getPlaylistConfig = {
+    url: "/api/user/playlists",
+    headers: { headers: { authorization: token } },
+    playlistId: playlistId,
+  };
+
+  useEffect(() => {
+    getPlayListFromServer(getPlaylistConfig);
+  }, [playlistId]);
 
   return (
     <div className="playlist-body">
